[stripe-ui-core] Move SectinoElement to stripe-ui-core
