+++ payments-ui-core/api/payments-ui-core.api
@@ -598,9 +598,6 @@ public final class com/stripe/android/ui/core/elements/SaveForFutureUseSpec$Comp
 	public final fun serializer ()Lkotlinx/serialization/KSerializer;
 }
 
-public final class com/stripe/android/ui/core/elements/SectionElementUIKt {
-}
-
 public final class com/stripe/android/ui/core/elements/SimpleDialogElementUIKt {
 }
