+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/UpiSpec.kt
@@ -3,6 +3,7 @@ package com.stripe.android.ui.core.elements
 import androidx.annotation.RestrictTo
 import com.stripe.android.ui.core.R
 import com.stripe.android.uicore.elements.IdentifierSpec
+import com.stripe.android.uicore.elements.SectionElement
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
