+++ stripe-ui-core/src/main/java/com/stripe/android/uicore/elements/SectionController.kt
@@ -1,10 +1,7 @@
-package com.stripe.android.ui.core.elements
+package com.stripe.android.uicore.elements
 
 import androidx.annotation.RestrictTo
 import androidx.annotation.StringRes
-import com.stripe.android.uicore.elements.Controller
-import com.stripe.android.uicore.elements.FieldError
-import com.stripe.android.uicore.elements.SectionFieldErrorController
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.combine
 
@@ -14,7 +11,7 @@ import kotlinx.coroutines.flow.combine
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 class SectionController(
     @StringRes val label: Int?,
-    val sectionFieldErrorControllers: List<SectionFieldErrorController>
+    sectionFieldErrorControllers: List<SectionFieldErrorController>
 ) : Controller {
     val error: Flow<FieldError?> = combine(
         sectionFieldErrorControllers.map {
